docs(build): add Renovate configuration details
Add Renovate configuration details for Azure DevOps pipelines.

This commit adds the necessary configuration details for Renovate to a `config.js` file, which supports NuGet, NPM, and GitHub authentication. Instructions for local debugging and an example configuration file are included, as well as an example YAML file for Renovate in Azure DevOps pipelines.

Note that the Renovatebot GitHub app is recommended for GitHub integrations as it provides smoother integration and checks automatically every 3 hours.